added app.tst.js
